# x-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a Paystack card payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiateInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiateResponse"
  /payments/stk-push:
    post:
      operationId: stkPush
      tags: [payments]
      summary: M-Pesa STK push via Paystack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StkPushInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpesaPushResult"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a Paystack payment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    PaymentInitiateInput:
      type: object
      required: [email, amount, planType]
      properties:
        email:
          type: string
        amount:
          type: integer
        planType:
          type: string
        currency:
          type: string
        callbackUrl:
          type: string
    PaymentInitiateResponse:
      type: object
      required: [authorizationUrl, reference, accessCode]
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
        accessCode:
          type: string
    StkPushInput:
      type: object
      required: [phone, amount, email]
      properties:
        phone:
          type: string
        amount:
          type: integer
        email:
          type: string
        planType:
          type: string
    MpesaPushResult:
      type: object
      required: [status, reference]
      properties:
        status:
          type: string
        reference:
          type: string
        message:
          type: string
    PaymentVerifyResponse:
      type: object
      required: [status, reference]
      properties:
        status:
          type: string
        reference:
          type: string
        amount:
          type: integer
        email:
          type: string
        paidAt:
          type: ["string", "null"]