# x-main/lib/api-zod/src/generated/types/mpesaPushResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */

export interface MpesaPushResult {
  status: string;
  reference: string;
  message?: string;
}