# x-main/lib/api-zod/src/generated/types/paymentVerifyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerifyResponse {
  status: string;
  reference: string;
  amount?: number;
  email?: string;
  /** @nullable */
  paidAt?: string | null;
}