# x-main/lib/api-zod/src/generated/types/stkPushInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */

export interface StkPushInput {
  phone: string;
  amount: number;
  email: string;
  planType?: string;
}